Edit scipt.js
removed the code that played the game 5 times.
Also removed any 'console.log' statements from the file.
Added new code for the manupilation the dom or the  UI elements inserted in 'index.html' so as to give the game a GUI interface rather than playing the game through the console.